+++ src/use_cases/comment/delete_comment.py
@@ -1,6 +1,13 @@
+from fastapi import HTTPException, status
+
+from exceptions.database import (
+    DataConflictError,
+    DatabaseError,
+    DatabaseUnavailableError,
+    ForeignKeyConflictError,
+)
 from infrastructure.postgres.repositories.comment_rep import (
     CommentRepository,)
-from exceptions.comment import CommentNotFound
 
 
 class DeleteCommentUseCase:
@@ -8,8 +15,36 @@ class DeleteCommentUseCase:
         self.repository = repository
 
     async def execute(self, comment_id: int) -> None:
+        if comment_id <= 0:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail="comment_id must be greater than 0",
+            )
         comment = await self.repository.get(comment_id)
         if comment is None:
-            raise CommentNotFound(comment_id)
-        
-        await self.repository.delete(comment)
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f"Comment with id={comment_id} not found",
+            )
+        try:
+            await self.repository.delete(comment)
+        except DataConflictError as err:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail=str(err),
+            ) from err
+        except ForeignKeyConflictError as err:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail=str(err),
+            ) from err
+        except DatabaseUnavailableError as err:
+            raise HTTPException(
+                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
+                detail=str(err),
+            ) from err
+        except DatabaseError as err:
+            raise HTTPException(
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+                detail=str(err),
+            ) from err
